requirements.txt
Django==1.9.7
git+https://github.com/erudit/erudit-core.git@master#egg=erudit-core
git+https://github.com/erudit/django-resumable-uploads.git@master#egg=django-resumable-uploads
git+https://github.com/erudit/django-account-actions.git@master#egg=django-account-actions
eulxml==1.0.1
eulfedora==1.5.0
-e git+https://github.com/erudit/liberuditarticle.git@master#egg=liberuditarticle
django-post-office==2.0.5
jsonfield==1.0.3
lxml==3.5.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==2.9.0
pluggy==0.3.0
mysqlclient==1.3.7
ptyprocess==0.5
py==1.4.30
pytz==2015.7
reportlab==3.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
PyMySQL==0.7.2
factory_boy==2.6.0
simplejson==3.8.2
django-filter==0.11.0
rules==1.1.1
PyPDF2==1.25.1
WeasyPrint==0.26
jsmin==2.2.0
django-modeltranslation==0.11
django-ckeditor==5.0.3
raven==5.10.2
sqlparse==0.1.18
django-ipware==1.1.3
feedparser==5.2.1
django-fsm==2.3.0
pysolr==3.4.0
fake-factory==0.5.3
gunicorn==19.4.5
html2text==2016.1.8
django-easy-pjax==1.3.0
celery==3.1.23
django-celery==3.1.17
django-js-reverse==0.7.1
djangorestframework==3.3.3
Markdown==2.6.6
django-widget-tweaks==1.4.1
Sickle==0.5
influxdb==2.12.0
